tests/wpt/tests/html/semantics/embedded-content/media-elements/interfaces/HTMLElement/HTMLTrackElement/readyState.html

<!doctype html>
<title>HTMLTrackElement.readyState</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function(){
    var track = document.createElement('track');
    assert_equals(track.readyState, 0);
}, document.title + ' default value');

test(function(){
    assert_equals(HTMLTrackElement.NONE, 0);
    assert_equals(HTMLTrackElement.LOADING, 1);
    assert_equals(HTMLTrackElement.LOADED, 2);
    assert_equals(HTMLTrackElement.ERROR, 3);
}, document.title + ' values');
</script>
